Fixed the ranking of old preference
Earlier for ranking we were adding the score of old preference to score based on filenames. Since we had ranking like lower score better rank.
It was decreasing the rank of user selected earlier choices.
Now the score from file name is divided by score of old prefrences, if old preferences is not there score is multipled by 2.