Rescanning of download page when folder is added
When new folder or the first time user adds permission, we need to rescan the entries in background page on the go to ensure selection is effective and can be choosen by user.
We call the background page and add call rescanning function when scan is started in the select page. Then at a interval of 10s we update the select with all the scanned values.