Continuing the scan in case of error(windows bug)
When we scan for the directories, In windows for some of the Galleries or file.We get the file modification error.
Due to this error scanning was stopping prematurely and we were not able to get to the point of scanning complete and show the visualization.

This should probably fix that. Need to be tested on windows.