Inital setup is launched and give permissions.
We launch the initial setup if the download directory given by the extension is not accessible or number of folders accessible are less than 8.

We ask user to give permission for more than 8 folders and download folder, if it is done we shutdown our initialsetup.
It will help user to improve experience of the app.